Move tests that now apply to Section and not LEDStrip to their own file
